backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for EstateAssess"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "redis>=5.0.1",
    "sqlalchemy>=2.0.27",
    "asyncpg>=0.29.0",
    "google-genai",
    "pydantic-settings>=2.2.1",
    "python-multipart>=0.0.9",
    "slowapi>=0.1.9",
    "reportlab>=4.1.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.28.1",
]

[tool.uv]
package = false
